Handle HTTP messages within the session manager.
